pyproject.toml
[project]
name = "grafana-cloud-demo-store"
version = "0.1.0"
description = "Interview-friendly ecommerce demo service instrumented for Grafana Cloud."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "opentelemetry-api>=1.38.0",
  "opentelemetry-sdk>=1.38.0",
  "opentelemetry-exporter-otlp-proto-http>=1.38.0",
  "opentelemetry-instrumentation-fastapi>=0.59b0",
  "opentelemetry-instrumentation-logging>=0.59b0",
  "uvicorn>=0.35.0",
]

[project.scripts]
demo-store = "app.main:run"
demo-traffic = "traffic:run"

[build-system]
requires = ["setuptools>=80"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["traffic"]

[tool.setuptools.packages.find]
include = ["app*"]
